requirements.txt
Pillow>=9.0
numpy>=1.20
pyfiglet>=0.8
